glibc: Turn off automatic requires
Dynamic requires is adding /bin/sh as a requires for the glibc
packages and subpackages. When we go to install glibc-devel into
the sysroot for a future build, the install fails because this
dynamic requires is not available (which is correct behavior).

Since we want to install but know we do not want to require
binaries in the sysroot environment (because we can't run them),
turn off dynamic requires.

This means we will need to explicitly Requires any dependent
packages instead of relying on the dynamic tooling.

Signed-off-by: Chris Co <chrco@microsoft.com>