container-host: Remove ExecStartPre from containerd service
Containerd.service failed to start because of a modprobe overlay
failure. If the kernel is built with the overlay module statically built
in, then this modprobe is expected to fail. So remove this ExecStartPre
command since current test kernels are all statically building the
overlay module into the kernel image.

Signed-off-by: Chris Co <chrco@microsoft.com>